AWSMobile Client fixes - Balance semaphore's wait and signal calls. (#1824)
GetIdentity inside AWSIdentityProvider.m process one request at a time. If a new request comes in while we are processing another, we ask the new request to wait using semaphore dispatch_semaphore_wait(self.semaphore, dispatch_time(DISPATCH_TIME_NOW, 5 * NSEC_PER_SEC));

This semaphore is signaled when we receive a request back for identity dispatch_semaphore_signal(self.semaphore);. But in the case where there is only one getIdentityId request, we signal the semaphore without calling a wait and thus increasing the semaphore value. So now when a new concurrent getIdentity is encountered, the semaphore is not decreased to a negative value and it wont wait.