Mark waiting upload parts as cancelled when canceling a multipart upload (#2288)
* Mark waiting upload parts as cancelled

An NSURLSession that has been marked as invalidated via `finishTasksAndInvalidate` will only fully invalidate when all tasks associated with it have been completed. Currently, only in-progress parts
are being cancelled and thus marked as completed. Any waiting tasks remain in the `NSURLSessionTaskStateSuspended` state. As per the documentation, the session will only be flagged as invalidated, calling
the `[NSURLSession session:didBecomeInvalidWithError:]` when all tasks have a state of `NSURLSessionTaskStateCompleted`.

* Added test to ensure session invalidation notification is received when removing a transfer utility