fix(AWSMobileClient): Federated SignIn completion handler is assigned before status change is issued
Federated sign in to Cognito Identity pool returns a status signedOutFederatedTokensInvalid when the token expire. When this happens,
a fetch to aws credential will wait till the user signin. But the user can break from this wait by calling releaseSignInWait. But
releaseSignInWait will only work if the completionHandler is set. This change make sure that the completion handler is set before the
signedOutFederatedTokensInvalid is published.