chore: Update Cognito ASF version
When we updated the license in the AWSCognitoIdentityProviderASF package, we
neglected to bump the version, which means that there is an inconsistency
between the podspec and what external tools like CocoaPods report. This change
rectifies that by bumping the version to 1.0.2, which should have been done
originally, and updating dependencies to use the new version. It's now Apache
all the way down.