fix(Cognito): Refresh Cognito user pool token if id token expired (#3220)
Description of changes: Cognito SDK does automatic token refresh only if the access token is expired. Now since Cognito added a new feature to change the expiry of id token as well, we need to check the validity of id token before returning from getSession call.

Also reduced the time window to check for expiry from 5 min to 2 min. This avoids refreshing the token more often if the token expiry is set to 5 min.