chore(AWSLex): Dropped support for arm64 iphonesimulator (#3365)
Drop the support for arm64 iphonesimulator in AWSLex
Updates the libAudioSourceiOS.a library to the latest (which has no support for arm64)

The arm64 iphonesimulator are the simulator in the new M1 mac systems. The internal libraries like libAudioSourceiOS.a do not currently have support for arm64 simulators. This PR is to temporary drop the support for arm64 iphonesimulator for AWSLex so that we can move forward with XCFramework support and SPM for the whole SDK.