WIP - cache path fix
* makes the stored property into a computed property

Since the URL for the Cache can change it is best to not hold onto it for a long living stored property. There may also be code which stores the full URL for the path which also need to be changed to a relative path on top of this computed property.